I1: part 1

+++ src/test/java/com/ibm/graph/client/TestSuite.java
@@ -29,27 +29,55 @@ public class TestSuite {
     @BeforeClass
     public static void setup() {
         Map envs = System.getenv();
-        TestSuite.graphClient = new IBMGraphClient(
-                envs.get("TEST_API_URL").toString(),
-                envs.get("TEST_USERNAME").toString(),
-                envs.get("TEST_PASSWORD").toString()
-        );
+
+        if((envs.get("TEST_API_URL") == null) || (envs.get("TEST_USERNAME") == null) || (envs.get("TEST_PASSWORD") == null)) {
+            logger.error("JUnit test aborted. Environment variables TEST_API_URL, TEST_USERNAME and TEST_PASSWORD must be set");
+            System.exit(1);
+        }
+
+        logger.info("******************************************************************************************************************");
+        if(logger.isDebugEnabled()) 
+            logger.info("DEBUG is enabled. To disable it, edit /src/test/resources/simplelogger.properties and follow the instructions.");
+        else 
+            logger.info("DEBUG is disabled. To enable it, edit /src/test/resources/simplelogger.properties and follow the instructions.");
+        logger.info("******************************************************************************************************************");
+
+        try {
+           TestSuite.graphClient = new IBMGraphClient(
+                                                        envs.get("TEST_API_URL").toString(),
+                                                        envs.get("TEST_USERNAME").toString(),
+                                                        envs.get("TEST_PASSWORD").toString());
+        }
+        catch(Exception ex) {
+            logger.error("Error creating IBMGraphClient", ex);   
+            TestSuite.graphClient = null;
+        }
     }
 
     @AfterClass
     public static void teardown() {
-        try {
-            logger.debug(String.format("Deleting graph with ID %s",graphId));
-            TestSuite.graphClient.deleteGraph(graphId);
+        if(TestSuite.graphClient != null) {
+           try {
+                logger.debug(String.format("Deleting graph with ID %s",graphId));
+                TestSuite.graphClient.deleteGraph(graphId);
+            }
+            catch(Exception ex) {
+                logger.error("Error deleting graph " + graphId + ": ", ex);
+            }
+            TestSuite.graphClient = null;
         }
-        catch(Exception ex) {
-            logger.error("Error deleting graph", ex);
-        }
-        TestSuite.graphClient = null;
     }
 
     public static void setGlobalGraphId(String graphId) {
-        TestSuite.graphClient.setGraph(graphId);
-        TestSuite.graphId = graphId;
+        if(TestSuite.graphClient != null) {
+           try {
+            TestSuite.graphClient.setGraph(graphId);
+            TestSuite.graphId = graphId;
+           }
+           catch(Exception ex) {
+                logger.error("Error switching to graph " + graphId + ": " , ex);
+           }
+        }
+
     }
 }
